Added CartesianRMQ implementation with missing Generator (full block queries are already fuctional)

+++ LogRMQ.cpp
@@ -7,21 +7,21 @@
 * 
 * @param length The length of the input vector.
 */
-int64_t calculateLayers(int64_t length) {
-	return ((int64_t)std::floor(std::log2(length))) + 1;
+uint64_t calculateLayers(uint64_t length) {
+	return ((uint64_t)std::floor(std::log2(length))) + 1;
 }
 
-std::pair<int64_t, int64_t> LogRMQ::accessLayerData(int64_t min, int64_t max) {
-	return savedAnswers->at(min)->at(std::log2(max - min + 1));
+std::pair<uint64_t, uint64_t> LogRMQ::accessLayerData(uint64_t min, uint64_t max) {
+	return savedAnswers->at(min)->at((uint64_t)std::log2(max - min + 1));
 }
 
-int64_t LogRMQ::rangeMinimumQuery(int64_t min, int64_t max) {
+uint64_t LogRMQ::rangeMinimumQuery(uint64_t min, uint64_t max) {
 	// All three query formulas as defined in the lecture.
-	int64_t l = (int64_t)std::floor(std::log2(max - min + 1));
-	int64_t splitMax = min + pow(2, l) - 1;
-	int64_t splitMin = max - pow(2, l) + 1;
-	std::pair<int64_t, int64_t> p1 = accessLayerData(min, splitMax);
-	std::pair<int64_t, int64_t> p2 = accessLayerData(splitMin, max);
+	uint64_t l = (uint64_t)std::floor(std::log2(max - min + 1));
+	uint64_t splitMax = (uint64_t)(min + pow(2, l) - 1);
+	uint64_t splitMin = (uint64_t)(max - pow(2, l) + 1);
+	std::pair<uint64_t, uint64_t> p1 = accessLayerData(min, splitMax);
+	std::pair<uint64_t, uint64_t> p2 = accessLayerData(splitMin, max);
 	if (p1.first <= p2.first) {
 		return p1.second; 
 	}
@@ -31,21 +31,21 @@ int64_t LogRMQ::rangeMinimumQuery(int64_t min, int64_t max) {
 }
 
 
-LogRMQ::LogRMQ(std::vector<int64_t> numbers) {
-	savedAnswers = new std::vector<std::vector<std::pair<int64_t, int64_t>>*>();
-	int64_t layer = calculateLayers(numbers.size());
+LogRMQ::LogRMQ(std::vector<uint64_t> numbers) {
+	savedAnswers = new std::vector<std::vector<std::pair<uint64_t, uint64_t>>*>();
+	uint64_t layer = calculateLayers(numbers.size());
 	// Initialize and fill in the first (trivial) values.
-	for (int64_t i = 0; i < numbers.size(); i++) {
-		savedAnswers->push_back(new std::vector<std::pair<int64_t, int64_t>>());
+	for (uint64_t i = 0; i < numbers.size(); i++) {
+		savedAnswers->push_back(new std::vector<std::pair<uint64_t, uint64_t>>());
 		savedAnswers->at(i)->push_back({numbers[i], i});
 	}
 	// Now fill in the rest. l and x are defined as in the lecture. We start with l = 1 because we already filled the trivial values.
-	for (int64_t l = 1; l <= layer; l++) {
-		for (int64_t x = 0; x < numbers.size(); x++) {
-			int64_t upperBound = pow(2, l) - 1 + x; // -1 because position x in inclusive.
+	for (uint64_t l = 1; l <= layer; l++) {
+		for (uint64_t x = 0; x < numbers.size(); x++) {
+			uint64_t upperBound = (uint64_t)(pow(2, l) - 1 + x); // -1 because position x in inclusive.
 			if (upperBound < numbers.size()) {
-				std::pair<int64_t, int64_t> p1 = savedAnswers->at(x)->at(l-1);
-				std::pair<int64_t, int64_t> p2 = savedAnswers->at(x + pow(2, l-1))->at(l-1); // Construction formula from the lecture
+				std::pair<uint64_t, uint64_t> p1 = savedAnswers->at(x)->at(l-1);
+				std::pair<uint64_t, uint64_t> p2 = savedAnswers->at((uint64_t)(x + pow(2, l-1)))->at(l-1); // Construction formula from the lecture
 				if (p1.first <= p2.first) { // p1 is smaller or same size
 					savedAnswers->at(x)->push_back(p1);
 				}
@@ -59,7 +59,7 @@ LogRMQ::LogRMQ(std::vector<int64_t> numbers) {
 
 
 LogRMQ::~LogRMQ() {
-	for (int64_t i = 0; i < savedAnswers->size(); i++) {
+	for (uint64_t i = 0; i < savedAnswers->size(); i++) {
 		delete(savedAnswers->at(i));
 	}
 	delete savedAnswers;
